configure esp32 to send post request to mongodb using spring

+++ backend-spring/src/main/java/org/example/backendspring/model/Pothole.java
@@ -12,6 +12,6 @@ import org.springframework.data.mongodb.core.mapping.Document;
 public class Pothole {
     private String class_name;
     private float probability;
-    private float inference_time_ms;
+    private int inference_time_ms;
     private String gps;
 }
